requirements.txt
scrapy>=2.11.0
playwright>=1.41.0
scrapy-playwright>=0.0.34
python-dotenv>=1.0.0 